Test(LottoTest) : Lotto 테스트 생성
